Revised the code structure and made corresponding changes.
