+++ src/pythermalcomfort/models.py
@@ -39,9 +39,6 @@ from pythermalcomfort.jos3_functions.construction import _to17array
 from pythermalcomfort.jos3_functions.parameters import Default, ALL_OUT_PARAMS
 from pythermalcomfort.__init__ import __version__
 
-# import functions for Zhang's sensation and comfort model
-from pythermalcomfort.zhang_comfort_model_functions import utilities, sensation, comfort
-
 
 def cooling_effect(tdb, tr, vr, rh, met, clo, wme=0, units="SI"):
     """Returns the value of the Cooling Effect (`CE`_) calculated in compliance
@@ -4645,674 +4642,4 @@ class JOS3:
         return self._version
 
 
-def zhang_sensation_comfort(
-    t_skin_local,
-    dt_skin_local_dt,
-    dt_core_local_dt,
-    t_skin_local_set={
-        "head": 35.3,
-        "neck": 35.6,
-        "chest": 35.1,
-        "back": 35.3,
-        "pelvis": 35.3,
-        "left_shoulder": 34.2,
-        "left_arm": 34.6,
-        "left_hand": 34.4,
-        "right_shoulder": 34.2,
-        "right_arm": 34.6,
-        "right_hand": 34.4,
-        "left_thigh": 34.3,
-        "left_leg": 32.8,
-        "left_foot": 33.3,
-        "right_thigh": 34.3,
-        "right_leg": 32.8,
-        "right_foot": 33.3,
-    },
-    options=None,
-):
-    """Zhang's local and overall sensation and comfort model.
 
-        This model predicts thermal sensation and comfort based on physiological responses,
-        such as skin temperature and its time variation.
-
-        It consists of four main helper functions that predict:
-        1. Local thermal sensation
-        2. Whole-body thermal sensation
-        3. Local thermal comfort
-        4. Whole-body thermal comfort
-
-        The results of these predictions are returned as a dictionary.
-
-        Accepted input data types for conditions like skin temperature are int, float, list, dict, and ndarray.
-        To avoid input errors, it's recommended to provide the data as a dictionary.
-        Each key in the dictionary should represent the name of a body segment, aligning with the JOS-3 model.
-
-        body_name_list = [
-            "head", "neck", "chest", "back", "pelvis",
-            "left_shoulder", "left_arm", "left_hand",
-            "right_shoulder", "right_arm", "right_hand",
-            "left_thigh", "left_leg", "left_foot",
-            "right_thigh", "right_leg", "right_foot"
-        ]
-
-        When data is provided as an int or float format, all body segments will be assigned the same value.
-        If data is provided as a list or ndarray format, it should be of length 17, with values corresponding to the keys listed above.
-
-
-        Parameters
-        -----------
-        t_skin_local : int, float, list, numpy.ndarray, dict
-            Local skin temperature [°C].
-        t_skin_local_set : int, float, list, numpy.ndarray, dict
-            Local skin temperature setpoint [°C].
-
-            Note: t_skin_local_set is the local skin temperature when people feel thermally neutral.
-            Neutral skin temperatures vary by site, generally higher for the trunk and lower for the extremities.
-            It is stated that multi-node segment model is used to get these parameters under the condition where
-            people thermally feel neutral, calculated by the PMV model at a certain metabolic rate and clothing level.
-
-            The experimental data on neutral skin temperature from Zhang's experiments : https://escholarship.org/uc/item/3f4599hx
-
-        dt_skin_local_dt : int, float, list, numpy.ndarray, dict
-            Time derivetive of local skin temperature [°C/s].
-        dt_core_local_dt : int, float, list, numpy.ndarray, dict
-            Time derivetive of core temperature [°C/s].
-        options : dict, optional
-            Dictionary containing options for the calculation.
-
-            Keys include:
-                "sensation_scale" (str):
-                    Scale for evaluating thermal sensation. Default is "9-point".
-                    Either '9-point' or '7-point'.
-                    9-point sensation scale (-4:very cold, -3:cold, -2:cool, -1:slightly cool, 0:neutral, +1:slightly warm, +2:warm, +3:hot, +4:very hot)
-                    7-point sensation scale (-3:cold, -2:cool, -1:slightly cool, 0:neutral, +1:slightly warm, +2:warm, +3:hot)
-
-                "is_transient" (bool):
-                    Indicates if the environment is transient. The original paper does not describe how to define.
-                    Default is False.
-
-                "is_controlled" (bool):
-                    Indicates if the environment is controlled. Default is False.
-
-                "show_extra_info" (bool):
-                    Indicates if extra information should be displayed. Default is False.
-
-        Returns
-        -------
-        sensation_local : local thermal sensation (each body segment) [-]
-        sensation_overall : overall thermal sensation [-]
-        comfort_local : local thermal comfort for each body segment [-]
-        comfort_overall : overall thermal sensation [-]
-        OPTIONAL PARAMETERS : the paramters listed below are returned if show_extra_info = True
-        t_skin_local : local skin temperature [°C].
-        t_skin_local_set : local skin temperature setpoint [°C].
-        dt_skin_local_dt : time derivetive of local skin temperature [°C/s].
-        t_skin_mean : mean skin temperature for the whole body [°C].
-        t_skin_mean_set : mean skin temperature setpoint for the whole body [°C].
-        dt_core_local_dt : time derivetive of core temperature [°C/s].
-
-
-        Examples:
-        ---------
-        >>> from pythermalcomfort.models import zhang_sensation_comfort
-        >>> dict_results = cbe_comfort(
-        >>>     t_skin_local={
-        >>>         "head": 34.3,
-        >>>         "neck": 34.6,
-        >>>         "chest": 34.1,
-        >>>         "back": 34.3,
-        >>>         "pelvis": 34.3,
-        >>>         "left_shoulder": 33.2,
-        >>>         "left_arm": 33.6,
-        >>>         "left_hand": 33.4,
-        >>>         "right_shoulder": 33.2,
-        >>>         "right_arm": 33.6,
-        >>>         "right_hand": 33.4,
-        >>>         "left_thigh": 33.3,
-        >>>         "left_leg": 31.8,
-        >>>         "left_foot": 32.3,
-        >>>         "right_thigh": 33.3,
-        >>>         "right_leg": 31.8,
-        >>>         "right_foot": 32.3,
-        >>>     },
-        >>>     t_skin_local_set={
-        >>>         "head": 35.3,
-        >>>         "neck": 35.6,
-        >>>         "chest": 35.1,
-        >>>         "back": 35.3,
-        >>>         "pelvis": 35.3,
-        >>>         "left_shoulder": 34.2,
-        >>>         "left_arm": 34.6,
-        >>>         "left_hand": 34.4,
-        >>>         "right_shoulder": 34.2,
-        >>>         "right_arm": 34.6,
-        >>>         "right_hand": 34.4,
-        >>>         "left_thigh": 34.3,
-        >>>         "left_leg": 32.8,
-        >>>         "left_foot": 33.3,
-        >>>         "right_thigh": 34.3,
-        >>>         "right_leg": 32.8,
-        >>>         "right_foot": 33.3,
-        >>>     },
-        >>>     dt_skin_local_dt={
-        >>>         "head": 0.01,
-        >>>         "neck": 0.01,
-        >>>         "chest": 0.01,
-        >>>         "back": 0.01,
-        >>>         "pelvis": 0.01,
-        >>>         "left_shoulder": 0.01,
-        >>>         "left_arm": 0.01,
-        >>>         "left_hand": 0.01,
-        >>>         "right_shoulder": 0.01,
-        >>>         "right_arm": 0.01,
-        >>>         "right_hand": 0.01,
-        >>>         "left_thigh": 0.01,
-        >>>         "left_leg": 0.01,
-        >>>         "left_foot": 0.01,
-        >>>         "right_thigh": 0.01,
-        >>>         "right_leg": 0.01,
-        >>>         "right_foot": 0.01,
-        >>>     },
-        >>>     dt_core_local_dt={
-        >>>         "head": 0.01,
-        >>>         "neck": 0.01,
-        >>>         "chest": 0.01,
-        >>>         "back": 0.01,
-        >>>         "pelvis": 0.01,
-        >>>         "left_shoulder": 0.01,
-        >>>         "left_arm": 0.01,
-        >>>         "left_hand": 0.01,
-        >>>         "right_shoulder": 0.01,
-        >>>         "right_arm": 0.01,
-        >>>         "right_hand": 0.01,
-        >>>         "left_thigh": 0.01,
-        >>>         "left_leg": 0.01,
-        >>>         "left_foot": 0.01,
-        >>>         "right_thigh": 0.01,
-        >>>         "right_leg": 0.01,
-        >>>         "right_foot": 0.01,
-        >>>     },
-        >>>     options={
-        >>>         "sensation_scale": "7-point",
-        >>>     }
-        >>> )
-        >>> print(dict_results)
-        {
-        'sensation_overall': -0.65,
-        'comfort_overall': -2.52,
-        'sensation_head': 0.15,
-        'sensation_neck': 0.15,
-        'sensation_chest': -3.0,
-        'sensation_back': -3.0,
-        'sensation_pelvis': -3.0,
-        'sensation_left_shoulder': 1.09,
-        'sensation_left_arm': 0.65,
-        'sensation_left_hand': 0.06,
-        'sensation_right_shoulder': 1.09,
-        'sensation_right_arm': 0.65,
-        'sensation_right_hand': 0.06,
-        'sensation_left_thigh': 2.23,
-        'sensation_left_leg': 1.54,
-        'sensation_left_foot': 1.12,
-        'sensation_right_thigh': 2.23,
-        'sensation_right_leg': 1.54,
-        'sensation_right_foot': 1.12,
-        'comfort_head': 2.53,
-        'comfort_neck': 1.98,
-        'comfort_chest': -2.42,
-        'comfort_back': -2.43,
-        'comfort_pelvis': -2.7,
-        'comfort_left_shoulder': 1.26,
-        'comfort_left_arm': 2.29,
-        'comfort_left_hand': 1.79,
-        'comfort_right_shoulder': 1.26,
-        'comfort_right_arm': 2.29,
-        'comfort_right_hand': 1.79,
-        'comfort_left_thigh': -1.36,
-        'comfort_left_leg': 0.5,
-        'comfort_left_foot': 1.56,
-        'comfort_right_thigh': -1.36,
-        'comfort_right_leg': 0.5,
-        'comfort_right_foot': 1.56
-    }
-    """
-    default_options = {
-        "sensation_scale": "9-point",
-        "is_transient": False,
-        "is_controlled": False,
-        "show_extra_info": False,
-    }
-
-    # If options is provided, update the default options with the provided ones
-    if options:
-        default_options.update(options)
-
-    options = default_options  # Now use the merged dictionary in your function
-
-    # Define phygiological input parameters
-    t_skin_local_array = utilities.convert_to_17_segments_array_from_any_data_type(
-        t_skin_local
-    )
-    dt_skin_local_dt_array = utilities.convert_to_17_segments_array_from_any_data_type(
-        dt_skin_local_dt
-    )
-    t_skin_local_set_array = utilities.convert_to_17_segments_array_from_any_data_type(
-        t_skin_local_set
-    )
-    t_skin_mean_value = np.average(t_skin_local_array, weights=Default.local_bsa)
-    t_skin_mean_set_value = np.average(
-        t_skin_local_set_array, weights=Default.local_bsa
-    )
-    dt_core_local_dt_array = utilities.convert_to_17_segments_array_from_any_data_type(
-        dt_core_local_dt
-    )
-
-    # Define optional input parameters
-    sensation_scale = options["sensation_scale"]
-    is_transient = options["is_transient"]
-    is_controlled = options["is_controlled"]
-    show_extra_info = options["show_extra_info"]
-
-    # ------------------------------------------------------
-    # Process Main
-    # ------------------------------------------------------
-
-    # Calculate local thermal sensation
-    sensation_local_array = sensation.calculate_local_thermal_sensation(
-        t_skin_local_array=t_skin_local_array,
-        dt_skin_local_dt_array=dt_skin_local_dt_array,
-        t_skin_local_set_array=t_skin_local_set_array,
-        t_skin_mean_value=t_skin_mean_value,
-        t_skin_mean_set_value=t_skin_mean_set_value,
-        dt_core_local_dt_array=dt_core_local_dt_array,
-        sensation_scale=sensation_scale,
-    )
-
-    # Calculate overall thermal sensation
-    overall_sensation_value = sensation.calculate_overall_sensation(
-        sensation_local_array=sensation_local_array
-    )
-
-    # Calculate local thermal comfort
-    comfort_local_array = comfort.calculate_local_thermal_comfort(
-        overall_sensation=overall_sensation_value,
-        sensation_local_array=sensation_local_array,
-    )
-
-    # Calculate overall thermal comfort
-    overall_comfort_value = comfort.calculate_overall_thermal_comfort(
-        local_comfort_array=comfort_local_array,
-        is_transient=is_transient,
-        is_controlled=is_controlled,
-    )
-
-    # Define result dict
-    dict_data = {
-        "sensation_local": np.around(sensation_local_array, 2),
-        "sensation_overall": np.around(overall_sensation_value, 2),
-        "comfort_local": np.around(comfort_local_array, 2),
-        "comfort_overall": np.around(overall_comfort_value, 2),
-    }
-
-    if show_extra_info:
-        dict_data.update(
-            {
-                "t_skin_local": t_skin_local_array,
-                "dt_skin_local_dt": dt_skin_local_dt_array,
-                "t_skin_local_set": t_skin_local_set_array,
-                "t_skin_mean_value": t_skin_mean_value,
-                "t_skin_mean_set_value": t_skin_mean_set_value,
-                "dt_core_local_dt_array": dt_core_local_dt_array,
-                "sensation_scale": sensation_scale,
-                "is_transient": is_transient,
-            }
-        )
-
-    # Create new dictionary to store the new keys and values
-    new_data_dict = {}
-
-    # Map 17 arrays of values to new keys for keys that contain 'local'
-    for key in dict_data.keys():
-        if "local" in key:
-            # Map 17 arrays of values to new keys
-            for i, body_name in enumerate(utilities.body_name_list):
-                # Replace 'local' with body_name in the key
-                new_key = key.replace("local", body_name)
-                new_value = dict_data[key][i]
-                new_data_dict[new_key] = new_value
-
-    # Merge the new dictionary with the original dictionary
-    dict_data.update(new_data_dict)
-
-    # Delete the original 'local_sensation' and 'local_comfort' keys
-    for key in list(dict_data.keys()):
-        if "local" in key and not any(part in key for part in utilities.body_name_list):
-            del dict_data[key]
-
-    return dict_data
-
-
-class ATCS(JOS3):
-    """ATCS (Advanced Thermal Comfort Simulator) is an integrated simulator that combines
-    JOS-3 model for simulating human thermal physiology, with a thermal sensation and
-    comfort model for predicting psychological parameters from the physiological parameters.
-    """
-
-    # Default skin temperature in degree Celsius for 17 local body parts
-    # The data comes from Hui Zhang's experiments
-    # https://escholarship.org/uc/item/3f4599hx
-    dict_default_t_skin = {
-        "head": 35.3,
-        "neck": 35.6,
-        "chest": 35.1,
-        "back": 35.3,
-        "pelvis": 35.3,
-        "left_shoulder": 34.2,
-        "left_arm": 34.6,
-        "left_hand": 34.4,
-        "right_shoulder": 34.2,
-        "right_arm": 34.6,
-        "right_hand": 34.4,
-        "left_thigh": 34.3,
-        "left_leg": 32.8,
-        "left_foot": 33.3,
-        "right_thigh": 34.3,
-        "right_leg": 32.8,
-        "right_foot": 33.3,
-    }
-
-    default_iteration_number_to_get_stable_condition = 600
-
-    def __init__(self, *args, **kwargs):
-        """Initializes this class by calling the parent class constructor
-        and setting initial values for instance variables.
-
-        Parameters
-        ----------
-        height : float, optional
-            body height, in [m]. The default is 1.72.
-        weight : float, optional
-            body weight, in [kg]. The default is 74.43.
-        fat : float, optional
-            fat percentage, in [%]. The default is 15.
-        age : int, optional
-            age, in [years]. The default is 20.
-        sex : str, optional
-            sex ("male" or "female"). The default is "male".
-        ci : float, optional
-            Cardiac index, in [L/min/m2]. The default is 2.6432.
-        bmr_equation : str, optional
-            The equation used to calculate basal metabolic rate (BMR). Choose a BMR equation.
-            The default is "harris-benedict" equation created uding Caucasian's data. (DOI: doi.org/10.1073/pnas.4.12.370)
-            If the Ganpule's equation (DOI: doi.org/10.1038/sj.ejcn.1602645) for Japanese people is used, input "japanese".
-        bsa_equation : str, optional
-            The equation used to calculate body surface area (bsa). Choose a bsa equation.
-            You can choose "dubois", "fujimoto", "kruazumi", or "takahira". The default is "dubois".
-            The body surface area can be calculated using the function
-            :py:meth:`pythermalcomfort.utilities.body_surface_area`.
-        ex_output : None, list or "all", optional
-            This is used when you want to display results other than the default output parameters (ex.skin temperature);
-            by default, JOS outputs only the most necessary parameters in order to reduce the computational load.
-            If the parameters other than the default output parameters are needed,
-            specify the list of the desired parameter names in string format like ["bf_skin", "bf_core", "t_artery"].
-            If you want to display all output results, set ex_output is "all".
-
-        Attributes
-        ----------
-        tdb : float or array-like
-            Dry bulb air temperature [°C].
-        tr : float or array-like
-            Mean radiant temperature [°C].
-        to : float or array-like
-            Operative temperature [°C].
-        v : float or array-like
-            Air speed [m/s].
-        rh : float or array-like
-            Relative humidity [%].
-        clo : float or array-like
-            Clothing insulation [clo].
-            Note: If you want to input clothing insulation to each body part,
-            it can be input using the dictionaly in "utilities.py" in "jos3_function" folder.
-            :py:meth:`pythermalcomfort.jos3_functions.utilities.local_clo_typical_ensembles`.
-        par : float
-            Physical activity ratio [-].
-            This equals the ratio of metabolic rate to basal metabolic rate.
-            The par of sitting quietly is 1.2.
-        posture : str
-            Body posture [-]. Choose a posture from standing, sitting or lying.
-        body_temp : numpy.ndarray (85,)
-            All segment temperatures of JOS-3
-
-        Methods
-        -------
-        simulate(times, dtime, output):
-            Run this model for given times.
-        dict_results():
-            Get results as a dictionary with pandas.DataFrame values.
-
-        Returns
-        -------
-        Predicted thermal phygiology
-            The same output as JOS-3 model
-        Predicted thermal sensation and comfort
-            Outputs from Zhang's sensation and comfort model with the inputs of JOS-3 model
-
-        """
-        # Initialize instance variables for this class.
-        self.cached_clo = np.ones(17) * Default.clothing_insulation
-        self.cached_par = Default.physical_activity_ratio
-        self.cached_overall_clo = np.average(self.cached_clo, weights=Default.local_bsa)
-
-        # Initialize parent class
-        self.comfort_setpt_skin = np.array(list(ATCS.dict_default_t_skin.values()))
-        self.t_skin_used = np.ones(17) * Default.skin_temperature
-        self.t_skin_wet = np.ones(17) * Default.skin_temperature
-        self.t_skin_previous_time_step = None
-        self.t_core_previous_time_step = None
-        self.overall_clo = Default.clothing_insulation
-
-        # Initialize parent class
-        super().__init__(*args, **kwargs)
-
-        self.overall_clo = np.average(self.clo, weights=Default.local_bsa)
-        print(f"During initialization, overall clo value is {self.overall_clo}")
-
-    def _calculate_comfort_setpt(self, overall_clo, par):
-        """Computes and returns the skin temperature when thermally neutral.
-
-        This method runs the JOS-3 model in uniform conditions at a temperature where PMV=0
-        for the provided metabolic rate and clothing insulation values. If the given values match
-        the cached ones, it directly returns the cached comfortable skin temperature set point.
-        Otherwise, it calculates and returns a new comfortable skin temperature set point.
-
-        Parameters
-        ----------
-        overall_clo : float
-            Overall clothing insulation value [clo]
-        par : float
-            Physical activity ratio [-]
-
-        Returns
-        -------
-        comfort_setpt_skin : np.array
-            Skin temperature set point for sensation/comfort model
-
-        Notes
-        -----
-        - If both `clo_overall` and `par` values match the cached values, this function will return
-          the cached comfortable skin temperature set point directly.
-        - The calculation of the comfortable skin temperature set point involves determining the operative temperature
-          where PMV is zero and then running the model in uniform conditions to extract skin temperatures as set points.
-        """
-        # If the clothing insulation value 'clo' and the metabolic rate 'par' are the same as the cached values,
-        # return the cached comfortable skin temperature set point
-        if overall_clo == self.cached_overall_clo and par == self.cached_par:
-            # print("Both 'clo' and 'par' are the same as the cached values. Cached values are applied to comfort set-points")
-            return self.comfort_setpt_skin
-
-        # Otherwise, calculate a new comfortable skin temperature set point
-        else:
-            # Calculate the metabolic rate based on the provided 'par' value.
-            min_met = 1.0
-            met = max(self.bmr * self.par / 58.15, min_met)
-
-            # Calculate the metabolic rate based on the provided 'par' value.
-            overall_clo = np.average(self._clo, weights=Default.local_bsa)
-
-            # Cash PAR (physical activity ratio) and overall clothing insulation
-            self.cached_par = self.par
-            self.cached_overall_clo = self.overall_clo
-
-            # Use the parent class's method to calculate the operative temperature where PMV is zero.
-            self.to = self._calculate_operative_temp_when_pmv_is_zero(
-                met=met, clo=overall_clo
-            )
-
-            # Run the model in uniform conditions at a temperature where PMV=0
-            # for the given met and clo and use the resulting skin temperatures as the setpoints
-            # Assuming that the initial conditions are uniform and at a temperature where PMV=0
-            print("Skin neutral temperature calculation starts")
-            print("Metabolic rate:", met)
-            print("Overall clo value:", overall_clo)
-            print("Operative temperature when PMV=0", self.to)
-            for i in range(10):
-                self._run(dtime=600, passive=False)
-
-            self.comfort_setpt_skin = self.t_skin
-            print("Comfort set-point temperature:", self.comfort_setpt_skin)
-
-            return self.comfort_setpt_skin
-
-    def _calculate_skin_temperature_considering_sweating(
-        self,
-        t_skin,
-        wet,
-    ):
-        """Calculate the skin temperature considering sweating for Zhang's model.
-
-        This function takes into account the impact of sweating on the skin temperature. The Zhang model's
-        prediction of local thermal sensation is represented by a logistic function of the difference between
-        the current skin temperature and the neutral skin temperature. In a hot environment, sweating tends to
-        suppress the increase in skin temperature, resulting in smaller differences in skin temperatures at
-        locations with higher skin temperatures (like the chest or back) under neutral thermal conditions. This
-        can lead to an underestimation of thermal sensations. Hence, this function modifies the skin temperature
-        based on skin wettedness, utilizing a simple consideration.
-        Based on the relationship between skin temperature and wettedness at isothermal sensation,
-        skin temperature is corrected by increasing skin wettedness to improve prediction accuracy at warm environments.
-
-        More details can be found in this Japanese research paper and data for TSV=3 is used as a reference:
-        https://eprints.lib.hokudai.ac.jp/dspace/bitstream/2115/7826/1/4-2-6_p67-72.pdf
-
-        Parameters
-        ----------
-        t_skin : float or numpy.ndarray
-            Skin temperature [°C]
-        wet : float
-            skin wettedness [-]
-
-        Returns
-        -------
-        t_skin_wet : float or numpy.ndarray
-            Corrected skin temperature, considering latent heat [°C]
-
-        Notes
-        -----
-        In the future, it may be more reasonable to predict thermal sensation
-        based on the total heat loss (sensible + latent) of JOS3 model, as in the SET calculation method.
-        """
-        coef = 2.5
-        t_skin_increase = coef * (wet - 0.06)
-        t_skin_wet = t_skin + t_skin_increase
-        return t_skin_wet
-
-    def _run(self, *args, **kwargs):
-        """Run the simulation for a single time step.
-
-        This method extends the `_run` method of the parent class. In addition to the parent's operations, it
-        calculates the rate of change of skin and core temperature and estimates thermal sensation and comfort parameters.
-
-        Parameters
-        ----------
-        *args :
-            Variable length argument list.
-        **kwargs :
-            Arbitrary keyword arguments.
-
-        Returns
-        -------
-        dict_data : dict
-            Dictionary containing data from the parent class `_run` method, rate of change of skin and core temperature,
-            and other thermal sensation and comfort parameters.
-
-        Notes
-        -----
-        - The skin temperature after considering sweat (`t_skin_wet`) is used for further calculations if `consider_sweat`
-          is True. Otherwise, the regular skin temperature (`t_skin`) is used.
-        - Local and overall thermal sensations and comfort are computed using the `cbe_comfort` function.
-        """
-
-        # Internal switch to consider sweat for comfort model
-        consider_sweat = True
-
-        # Define comfort set-point temperature
-        comfort_setpt_skin = self._calculate_comfort_setpt(
-            overall_clo=self.overall_clo, par=self.par
-        )
-        # Calculate the skin temperature after accounting for the effects of sweating.
-        self.t_skin_wet = self._calculate_skin_temperature_considering_sweating(
-            t_skin=self.t_skin, wet=self.w
-        )
-
-        # Use t_skin_wet for calculations if consider_sweat is True
-        self.t_skin_used = self.t_skin_wet if consider_sweat else self.t_skin
-
-        # Execute the `_run` method from the parent class and store its returned data.
-        dict_data = super()._run(*args, **kwargs)
-
-        # Retrieve or default the time difference value from keyword arguments.
-        dtime = kwargs.get("dtime", 60)
-
-        # Calculate the rate of change of skin and core temperatures over time.
-        dt_skin = (
-            (self.t_skin_used - self.t_skin_previous_time_step) / dtime
-            if self.t_skin_previous_time_step is not None
-            else 0
-        )
-        dt_core = (
-            (self.t_core - self.t_core_previous_time_step) / dtime
-            if self.t_core_previous_time_step is not None
-            else 0
-        )
-
-        # Store the used skin and core temperatures for future calculations.
-        self.t_skin_previous_time_step = self.t_skin_used
-        self.t_core_previous_time_step = self.t_core
-
-        # Update the dictionary with the rate of change values for skin and core temperatures.
-        dict_data.update(
-            {
-                "dt_skin": dt_skin,
-                "dt_core": dt_core,
-            }
-        )
-
-        # Calculate thermal sensations and comfort parameters
-        dict_sensation_and_comfort = zhang_sensation_comfort(
-            t_skin_local=self.t_skin_used,
-            dt_skin_local_dt=dt_skin,
-            t_skin_local_set=comfort_setpt_skin,
-            dt_core_local_dt=dt_core,
-        )
-
-        # Update the dictionary with the thermal sensation and comfort parameters
-        dict_data.update(dict_sensation_and_comfort)
-
-        return dict_data
-
-    def simulate(self, *args, **kwargs):
-        """Runs the model for a given time.
-        This method extends the simulate method of the parent class by additionally computing the comfortable skin
-        temperature set point before running the simulation.
-        """
-        result = super().simulate(*args, **kwargs)
-        return result
