--- src/pythermalcomfort/zhang_comfort_model_functions/coefficients.py
@@ -1,421 +1,0 @@
-import pandas as pd  # for making dataframe
-
-c1_plus = {
-    "head": 0.9,
-    "neck": 0.9,
-    "chest": 1.0,
-    "back": 1.0,
-    "pelvis": 0.4,
-    "left_shoulder": 0.4,
-    "left_arm": 0.7,
-    "left_hand": 0.45,
-    "right_shoulder": 0.4,
-    "right_arm": 0.7,
-    "right_hand": 0.45,
-    "left_thigh": 0.29,
-    "left_leg": 0.4,
-    "left_foot": 0.26,
-    "right_thigh": 0.29,
-    "right_leg": 0.4,
-    "right_foot": 0.26,
-}
-c1_minus = {
-    "head": 0.38,
-    "neck": 0.38,
-    "chest": 0.35,
-    "back": 0.3,
-    "pelvis": 0.2,
-    "left_shoulder": 0.29,
-    "left_arm": 0.3,
-    "left_hand": 0.2,
-    "right_shoulder": 0.29,
-    "right_arm": 0.3,
-    "right_hand": 0.2,
-    "left_thigh": 0.2,
-    "left_leg": 0.29,
-    "left_foot": 0.25,
-    "right_thigh": 0.2,
-    "right_leg": 0.29,
-    "right_foot": 0.25,
-}
-
-k1 = {
-    "head": 0.18,
-    "neck": 0.18,
-    "chest": 0.1,
-    "back": 0.1,
-    "pelvis": 0.15,
-    "left_shoulder": 0.1,
-    "left_arm": 0.1,
-    "left_hand": 0.15,
-    "right_shoulder": 0.1,
-    "right_arm": 0.1,
-    "right_hand": 0.15,
-    "left_thigh": 0.11,
-    "left_leg": 0.1,
-    "left_foot": 0.15,
-    "right_thigh": 0.11,
-    "right_leg": 0.1,
-    "right_foot": 0.15,
-}
-
-c2_plus = {
-    "head": 90,
-    "neck": 90,
-    "chest": 136,
-    "back": 192,
-    "pelvis": 137,
-    "left_shoulder": 167,
-    "left_arm": 125,
-    "left_hand": 46,
-    "right_shoulder": 167,
-    "right_arm": 125,
-    "right_hand": 46,
-    "left_thigh": 263,
-    "left_leg": 212,
-    "left_foot": 162,
-    "right_thigh": 263,
-    "right_leg": 212,
-    "right_foot": 162,
-}
-
-c2_minus = {
-    "head": 543,
-    "neck": 543,
-    "chest": 39,
-    "back": 88,
-    "pelvis": 75,
-    "left_shoulder": 156,
-    "left_arm": 144,
-    "left_hand": 19,
-    "right_shoulder": 156,
-    "right_arm": 144,
-    "right_hand": 19,
-    "left_thigh": 151,
-    "left_leg": 206,
-    "left_foot": 109,
-    "right_thigh": 151,
-    "right_leg": 206,
-    "right_foot": 109,
-}
-
-c3 = {
-    "head": 0,
-    "neck": 0,
-    "chest": -2135,
-    "back": -4054,
-    "pelvis": -5053,
-    "left_shoulder": 0,
-    "left_arm": 0,
-    "left_hand": 0,
-    "right_shoulder": 0,
-    "right_arm": 0,
-    "right_hand": 0,
-    "left_thigh": 0,
-    "left_leg": 0,
-    "left_foot": 0,
-    "right_thigh": 0,
-    "right_leg": 0,
-    "right_foot": 0,
-}
-
-a_if_delta_sens_local_is_less_than_minus_2 = {
-    "head": 0.54,
-    "neck": 0.65,
-    "chest": 0.91,
-    "back": 0.91,
-    "pelvis": 0.94,
-    "left_shoulder": 0.43,
-    "left_arm": 0.37,
-    "left_hand": 0.25,
-    "right_shoulder": 0.43,
-    "right_arm": 0.37,
-    "right_hand": 0.25,
-    "left_thigh": 0.81,
-    "left_leg": 0.7,
-    "left_foot": 0.5,
-    "right_thigh": 0.81,
-    "right_leg": 0.7,
-    "right_foot": 0.5,
-}
-
-a_if_delta_sens_local_is_between_minus_2_to_plus_2 = {
-    "head": 0.5,
-    "neck": 0.46,
-    "chest": 0.57,
-    "back": 0.46,
-    "pelvis": 0.32,
-    "left_shoulder": 0.28,
-    "left_arm": 0.38,
-    "left_hand": 0.0,
-    "right_shoulder": 0.28,
-    "right_arm": 0.38,
-    "right_hand": 0.0,
-    "left_thigh": 0.3,
-    "left_leg": 0.29,
-    "left_foot": 0.0,
-    "right_thigh": 0.3,
-    "right_leg": 0.29,
-    "right_foot": 0.0,
-}
-
-a_if_delta_sens_local_is_plus_2_or_more = {
-    "head": 0.4,
-    "neck": 0.4,
-    "chest": 0.4,
-    "back": 0.2,
-    "pelvis": 0.6,
-    "left_shoulder": 0.4,
-    "left_arm": 0.3,
-    "left_hand": 0.1,
-    "right_shoulder": 0.4,
-    "right_arm": 0.3,
-    "right_hand": 0.1,
-    "left_thigh": 0.4,
-    "left_leg": 0.4,
-    "left_foot": 0.6,
-    "right_thigh": 0.4,
-    "right_leg": 0.4,
-    "right_foot": 0.6,
-}
-
-b_if_delta_sens_local_is_less_than_minus_2 = {
-    "head": -1.1,
-    "neck": -0.92,
-    "chest": -1.14,
-    "back": -0.92,
-    "pelvis": -0.64,
-    "left_shoulder": -0.56,
-    "left_arm": -0.73,
-    "left_hand": 0.0,
-    "right_shoulder": -0.56,
-    "right_arm": -0.73,
-    "right_hand": 0.0,
-    "left_thigh": -0.6,
-    "left_leg": -0.59,
-    "left_foot": 0.0,
-    "right_thigh": -0.6,
-    "right_leg": -0.59,
-    "right_foot": 0.0,
-}
-
-b_if_delta_sens_local_is_between_minus_2_to_plus_2 = {
-    "head": 0.0,
-    "neck": 0.0,
-    "chest": 0.0,
-    "back": 0.0,
-    "pelvis": 0.0,
-    "left_shoulder": 0.0,
-    "left_arm": 0.0,
-    "left_hand": 0.0,
-    "right_shoulder": 0.0,
-    "right_arm": 0.0,
-    "right_hand": 0.0,
-    "left_thigh": 0.0,
-    "left_leg": 0.0,
-    "left_foot": 0.0,
-    "right_thigh": 0.0,
-    "right_leg": 0.0,
-    "right_foot": 0.0,
-}
-
-b_if_delta_sens_local_is_plus_2_or_more = {
-    "head": 1.1,
-    "neck": 0.63,
-    "chest": 1.14,
-    "back": 0.92,
-    "pelvis": 0.64,
-    "left_shoulder": 0.56,
-    "left_arm": 0.73,
-    "left_hand": 0.0,
-    "right_shoulder": 0.56,
-    "right_arm": 0.73,
-    "right_hand": 0.0,
-    "left_thigh": 0.6,
-    "left_leg": 0.59,
-    "left_foot": 0.0,
-    "right_thigh": 0.6,
-    "right_leg": 0.59,
-    "right_foot": 0.0,
-}
-
-c_if_delta_sens_local_is_less_than_minus_2 = {
-    "head": -2,
-    "neck": -2,
-    "chest": -2,
-    "back": -2,
-    "pelvis": -2,
-    "left_shoulder": -2,
-    "left_arm": -2,
-    "left_hand": -2,
-    "right_shoulder": -2,
-    "right_arm": -2,
-    "right_hand": -2,
-    "left_thigh": -2,
-    "left_leg": -2,
-    "left_foot": -2,
-    "right_thigh": -2,
-    "right_leg": -2,
-    "right_foot": -2,
-}
-
-c_if_delta_sens_local_is_between_minus_2_to_plus_2 = {
-    "head": 0.0,
-    "neck": 0.0,
-    "chest": 0.0,
-    "back": 0.0,
-    "pelvis": 0.0,
-    "left_shoulder": 0.0,
-    "left_arm": 0.0,
-    "left_hand": 0.0,
-    "right_shoulder": 0.0,
-    "right_arm": 0.0,
-    "right_hand": 0.0,
-    "left_thigh": 0.0,
-    "left_leg": 0.0,
-    "left_foot": 0.0,
-    "right_thigh": 0.0,
-    "right_leg": 0.0,
-    "right_foot": 0.0,
-}
-
-c_if_delta_sens_local_is_plus_2_or_more = {
-    "head": 2,
-    "neck": 2,
-    "chest": 2,
-    "back": 2,
-    "pelvis": 2,
-    "left_shoulder": 2,
-    "left_arm": 2,
-    "left_hand": 2,
-    "right_shoulder": 2,
-    "right_arm": 2,
-    "right_hand": 2,
-    "left_thigh": 2,
-    "left_leg": 2,
-    "left_foot": 2,
-    "right_thigh": 2,
-    "right_leg": 2,
-    "right_foot": 2,
-}
-
-C31 = {
-    "head": -0.35,
-    "neck": 0.0,
-    "chest": -0.66,
-    "back": -0.45,
-    "pelvis": -0.59,
-    "left_shoulder": -0.3,
-    "left_arm": -0.23,
-    "left_hand": -0.8,
-    "right_shoulder": -0.3,
-    "right_arm": -0.23,
-    "right_hand": -0.8,
-    "left_thigh": 0.0,
-    "left_leg": -0.2,
-    "left_foot": -0.91,
-    "right_thigh": 0.0,
-    "right_leg": -0.2,
-    "right_foot": -0.91,
-}
-
-C32 = {
-    "head": 0.35,
-    "neck": 0.0,
-    "chest": 0.66,
-    "back": 0.45,
-    "pelvis": 0.0,
-    "left_shoulder": 0.35,
-    "left_arm": 0.23,
-    "left_hand": 0.8,
-    "right_shoulder": 0.35,
-    "right_arm": 0.23,
-    "right_hand": 0.8,
-    "left_thigh": 0.0,
-    "left_leg": 0.61,
-    "left_foot": 0.4,
-    "right_thigh": 0.0,
-    "right_leg": 0.61,
-    "right_foot": 0.4,
-}
-
-C6 = {
-    "head": 2.17,
-    "neck": 1.96,
-    "chest": 2.1,
-    "back": 2.1,
-    "pelvis": 2.06,
-    "left_shoulder": 2.14,
-    "left_arm": 2.0,
-    "left_hand": 1.98,
-    "right_shoulder": 2.14,
-    "right_arm": 2.0,
-    "right_hand": 1.98,
-    "left_thigh": 1.98,
-    "left_leg": 2.0,
-    "left_foot": 2.13,
-    "right_thigh": 1.98,
-    "right_leg": 2.0,
-    "right_foot": 2.13,
-}
-
-C71 = {
-    "head": 0.28,
-    "neck": 0.0,
-    "chest": 1.39,
-    "back": 0.96,
-    "pelvis": 0.5,
-    "left_shoulder": 0.0,
-    "left_arm": 0.0,
-    "left_hand": 0.48,
-    "right_shoulder": 0.0,
-    "right_arm": 0.0,
-    "right_hand": 0.48,
-    "left_thigh": 0.0,
-    "left_leg": 1.67,
-    "left_foot": 0.5,
-    "right_thigh": 0.0,
-    "right_leg": 1.67,
-    "right_foot": 0.5,
-}
-
-C72 = {
-    "head": 0.4,
-    "neck": 0.0,
-    "chest": 0.9,
-    "back": 0.0,
-    "pelvis": 0.0,
-    "left_shoulder": 0.0,
-    "left_arm": 1.71,
-    "left_hand": 0.48,
-    "right_shoulder": 0.0,
-    "right_arm": 1.71,
-    "right_hand": 0.48,
-    "left_thigh": 0.0,
-    "left_leg": 0.0,
-    "left_foot": 0.3,
-    "right_thigh": 0.0,
-    "right_leg": 0.0,
-    "right_foot": 0.3,
-}
-
-C8 = {
-    "head": 0.5,
-    "neck": -0.19,
-    "chest": 0.0,
-    "back": 0.0,
-    "pelvis": -0.51,
-    "left_shoulder": -0.4,
-    "left_arm": -0.68,
-    "left_hand": 0.0,
-    "right_shoulder": -0.4,
-    "right_arm": -0.68,
-    "right_hand": 0.0,
-    "left_thigh": 0.0,
-    "left_leg": 0.0,
-    "left_foot": 0.0,
-    "right_thigh": 0.0,
-    "right_leg": 0.0,
-    "right_foot": 0.0,
-}
-
