--- src/pythermalcomfort/zhang_comfort_model_functions/coefficients_summary.csv
@@ -1,18 +1,0 @@
-,c1_plus,c1_minus,k1,c2_plus,c2_minus,c3,a_if_delta_sens_local_is_less_than_minus_2,a_if_delta_sens_local_is_between_minus_2_to_plus_2,a_if_delta_sens_local_is_plus_2_or_more,b_if_delta_sens_local_is_less_than_minus_2,b_if_delta_sens_local_is_between_minus_2_to_plus_2,b_if_delta_sens_local_is_plus_2_or_more,c_if_delta_sens_local_is_less_than_minus_2,c_if_delta_sens_local_is_between_minus_2_to_plus_2,c_if_delta_sens_local_is_plus_2_or_more,C31,C32,C6,C71,C72,C8
-head,0.9,0.38,0.18,90,543,0,0.54,0.5,0.4,-1.1,0.0,1.1,-2,0.0,2,-0.35,0.35,2.17,0.28,0.4,0.5
-neck,0.9,0.38,0.18,90,543,0,0.65,0.46,0.4,-0.92,0.0,0.63,-2,0.0,2,0.0,0.0,1.96,0.0,0.0,-0.19
-chest,1.0,0.35,0.1,136,39,-2135,0.91,0.57,0.4,-1.14,0.0,1.14,-2,0.0,2,-0.66,0.66,2.1,1.39,0.9,0.0
-back,1.0,0.3,0.1,192,88,-4054,0.91,0.46,0.2,-0.92,0.0,0.92,-2,0.0,2,-0.45,0.45,2.1,0.96,0.0,0.0
-pelvis,0.4,0.2,0.15,137,75,-5053,0.94,0.32,0.6,-0.64,0.0,0.64,-2,0.0,2,-0.59,0.0,2.06,0.5,0.0,-0.51
-left_shoulder,0.4,0.29,0.1,167,156,0,0.43,0.28,0.4,-0.56,0.0,0.56,-2,0.0,2,-0.3,0.35,2.14,0.0,0.0,-0.4
-left_arm,0.7,0.3,0.1,125,144,0,0.37,0.38,0.3,-0.73,0.0,0.73,-2,0.0,2,-0.23,0.23,2.0,0.0,1.71,-0.68
-left_hand,0.45,0.2,0.15,46,19,0,0.25,0.0,0.1,0.0,0.0,0.0,-2,0.0,2,-0.8,0.8,1.98,0.48,0.48,0.0
-right_shoulder,0.4,0.29,0.1,167,156,0,0.43,0.28,0.4,-0.56,0.0,0.56,-2,0.0,2,-0.3,0.35,2.14,0.0,0.0,-0.4
-right_arm,0.7,0.3,0.1,125,144,0,0.37,0.38,0.3,-0.73,0.0,0.73,-2,0.0,2,-0.23,0.23,2.0,0.0,1.71,-0.68
-right_hand,0.45,0.2,0.15,46,19,0,0.25,0.0,0.1,0.0,0.0,0.0,-2,0.0,2,-0.8,0.8,1.98,0.48,0.48,0.0
-left_thigh,0.29,0.2,0.11,263,151,0,0.81,0.3,0.4,-0.6,0.0,0.6,-2,0.0,2,0.0,0.0,1.98,0.0,0.0,0.0
-left_leg,0.4,0.29,0.1,212,206,0,0.7,0.29,0.4,-0.59,0.0,0.59,-2,0.0,2,-0.2,0.61,2.0,1.67,0.0,0.0
-left_foot,0.26,0.25,0.15,162,109,0,0.5,0.0,0.6,0.0,0.0,0.0,-2,0.0,2,-0.91,0.4,2.13,0.5,0.3,0.0
-right_thigh,0.29,0.2,0.11,263,151,0,0.81,0.3,0.4,-0.6,0.0,0.6,-2,0.0,2,0.0,0.0,1.98,0.0,0.0,0.0
-right_leg,0.4,0.29,0.1,212,206,0,0.7,0.29,0.4,-0.59,0.0,0.59,-2,0.0,2,-0.2,0.61,2.0,1.67,0.0,0.0
-right_foot,0.26,0.25,0.15,162,109,0,0.5,0.0,0.6,0.0,0.0,0.0,-2,0.0,2,-0.91,0.4,2.13,0.5,0.3,0.0
